pyproject.toml
[tool.poetry]
name = "stupice"
version = "0.1.0"
description = "Another Pomodoro App"
authors = ["Pierre-Yves Gillier <bonjour@pygillier.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5.1"
pygobject = "^3.36.1"
python-statemachine = "^0.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
app = "stupice:main"
